+++ src/onegov.election_day/onegov/election_day/views/auth.py
@@ -80,10 +80,10 @@ def handle_password_reset_request(self, request, form):
                 request.link(self, name='reset-password')
             )
 
-            request.app.send_email(
+            request.app.send_transactional_email(
                 subject=request.translate(_("Password reset")),
                 receivers=(user.username, ),
-                reply_to=request.app.mail_sender,
+                reply_to=request.app.mail['transactional']['sender'],
                 content=render_template(
                     'mail_password_reset.pt',
                     request,