+++ src/onegov.election_day/HISTORY.rst
@@ -4,6 +4,9 @@ Changelog
 - Optimizes sending email notifications.
   [msom]
 
+- Splits e-mails into transactional/marketing.
+  [href]
+
 2.1.0 (2018-03-05)
 ~~~~~~~~~~~~~~~~~~~